Reverse SectionFooter/Header(see PhpWord\Element\)
